Move SchemaOptions out of components folder because they are not components
